create install structure if not exist

+++ avx.bat
@@ -22,10 +22,13 @@ if %1 == 0 (
 	echo ~~-~~ SSE2 ~~-~~
 )
 
-rmdir /S /Q %PATH_INSTALL%
+if exist %PATH_INSTALL%\. rmdir /S /Q %PATH_INSTALL%
 mklink /J %PATH_INSTALL% %PATH_HTTPD_SDK%\release\%MSVC_DEPS%_%ARCH%%AVXB%
+if not exist %PATH_INSTALL%\lib\. mkdir %PATH_INSTALL%\lib
+if not exist %PATH_INSTALL%\bin\. mkdir %PATH_INSTALL%\bin
+if not exist %PATH_INSTALL%\include\. mkdir %PATH_INSTALL%\include
 
-REM /GL ko sur nghttp2. /GL ko si /Zi http://forums.codeguru.com/showthread.php?556135-ZI-and-GL-incompatible
+	REM /GL ko sur nghttp2. /GL ko si /Zi http://forums.codeguru.com/showthread.php?556135-ZI-and-GL-incompatible
 set EXTCFLAGSNGHTTP2=/GS- /Oy- /guard:cf- /FD /GF /Zc:inline /MP%NUMBER_OF_PROCESSORS% /LD /MD /Zi /Ox %AVX%
 set EXTCFLAGS=/GL %EXTCFLAGSNGHTTP2%
